dags/requirements.txt
# stock_pipeline/dags/requirements.txt
apache-airflow-providers-postgres
